+++ internal/data/ent/schema/spu_img.go
@@ -19,7 +19,7 @@ func (SpuImg) Annotations() []schema.Annotation {
 func (SpuImg) Fields() []ent.Field {
 	return []ent.Field{
 		field.String("img").Comment(""),
-		field.Int("spu_id").Comment(""),
+		field.Int64("spu_id").Comment(""),
 	}
 }
 func (SpuImg) Mixin() []ent.Mixin {
